--- a/src/models/index.ts
+++ b/src/models/index.ts
@@ -13,3 +13,4 @@ export * from "./WCSType";
 export * from "./RegionCreationMode";
 export * from "./CompressionQuality";
 export * from "./TileCache";
+export * from "./Processed";