… src/models/Processed.ts
… src/services/BackendService.ts
… src/stores/AppStore.ts
… src/stores/FrameStore.ts
… src/stores/SpatialProfileStore.ts
… src/stores/SpectralProfileStore.ts