Note: Companion PR of [#262](https://github.com/CARTAvis/carta-backend/pull/262). This PR should not be merged until the companion PR is approved.

This PR adds decimation to spatial profiles, to ensure that performance is smooth when viewing large images. See #366 for details and some benchmarks. 

This PR also adapts the backend to the changes in ICD version 5.0.0 (described in the google doc). Specifically, it makes use of `bytes` fields instead of `repeated float` and `repeated double` for spatial and spectral profiles. This is done to avoid the overhead incurred when decoding these messages, particularly in the case when the profiles are very large. This overhead was up to 2 ms for large (48K profiles). The decode time has now decreased to below 0.25 ms. 

Finally, it makes use of the `fileFeatureFlags` value returned by `OPEN_FILE_ACK` to determine whether swizzled datasets are present, and then adjusts the throttled cursor update performance accordingly. This should enable _up to_ 5 fps spatial and spectral profile updates for FITS/CASA files, and up to 10 fps for HDF5 files.